+++ src/index.ts
@@ -30,3 +30,6 @@ export { default as PositionManager, createPositionManagerConfig, initializeConf
 export type { PositionManagerConfig } from './core/PositionManager.js';
 
 export { log, warn } from './utils/logger.js';
+
+export { default as ReferencePlace, NO_REFERENCE_PLACE, VALID_REF_PLACE_CLASSES } from './core/ReferencePlace.js';
+export type { OsmElement } from './core/ReferencePlace.js';